+++ changelog.md
@@ -20,12 +20,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `added` Meta SEO files `gh-pages` branch to `src`.
   - `chngd` Build folder from `dist/albatroz` to `dist`.
 
-### [mc8] - 08-02-2019 00:23
+### [mc8] - 08-02-2019 00:23 57aca14ec82a036626f0ad18559168ddc21801ad
   - `added` "@angular/flex-layout"
   - `added` Flexible menu toolbar.
 
-### [mc9] - 08-02-2019
+### [mc9] - 08-02-2019 17:34 3276d0a66ba9d22b98cf248c8da1298236195414
   - `added` SharedModule with `header` and `footer` layouts.
   - `chngd` Navbar loading throug header component.
   - `chngd` Prebuild theme to `indigo-pink`.
   - `fixed` Menu icon by importing Material Icons Font.
+
+### [mc10] - 19-02-2019 
+  - `added` Swith to Material Angular Landing Theme.
+  - `added` Main page.
+  - `chngd` Update all modules to Angular 7.
+  - `chngd` Template engine from SASS to SCSS.
+  - `fixed` Opitimase all gallery images size.